transactional archivation works

+++ src/main/java/ua/kpi/tef/zu/gp3servlet/repository/GenericDao.java
@@ -8,11 +8,11 @@ import java.util.Optional;
 /**
  * Created by Anton Domin on 2020-04-16
  */
-public interface GenericDao<T> extends AutoCloseable{
+public interface GenericDao<T> extends AutoCloseable {
 	void create (T entity) throws DatabaseException;
 	Optional<T> findById(long id) throws DatabaseException;
 	List<T> findAll() throws DatabaseException;
 	void update(T entity) throws DatabaseException;
-	void delete(int id) throws DatabaseException;
+	void delete(long id) throws DatabaseException;
 }
